rename service for operative and calculation tag names
